+++ packages/bloc_test/example/main.dart
@@ -5,10 +5,11 @@ import 'package:bloc_test/bloc_test.dart';
 import 'package:test/test.dart';
 
 // Mock Cubit
-class MockCounterCubit extends MockBloc<int> implements CounterCubit {}
+class MockCounterCubit extends MockCubit<int> implements CounterCubit {}
 
 // Mock Bloc
-class MockCounterBloc extends MockBloc<int> implements CounterBloc {}
+class MockCounterBloc extends MockBloc<CounterEvent, int>
+    implements CounterBloc {}
 
 void main() {
   mainCubit();
@@ -40,7 +41,7 @@ void mainCubit() {
     blocTest<CounterCubit, int>(
       'emits [1] when increment is called',
       build: () => CounterCubit(),
-      act: (cubit) async => cubit.increment(),
+      act: (cubit) => cubit.increment(),
       expect: const <int>[1],
     );
   });
@@ -71,7 +72,7 @@ void mainBloc() {
     blocTest<CounterBloc, int>(
       'emits [1] when CounterEvent.increment is added',
       build: () => CounterBloc(),
-      act: (bloc) async => bloc.add(CounterEvent.increment),
+      act: (bloc) => bloc.add(CounterEvent.increment),
       expect: const <int>[1],
     );
   });