+++ packages/bloc_test/lib/bloc_test.dart
@@ -1,6 +1,5 @@
 library bloc_test;
 
 export 'src/bloc_test.dart';
-export 'src/emits_exactly.dart';
 export 'src/mock_bloc.dart';
 export 'src/when_listen.dart';